So.. there is some more changes.
We have a default Server. This is for any user and any camera. We want to have a default Org Server too. So users within those orgs can by default use org default servers . In case org default server is not available, then use the default server. So we need to organise the servers collection accordingly. Can you fix it