isDefault can be isSystemDefault.
Ideally, when a new org is created, thats when it takes the default system server as the org default server.

later on, the system admin, may or maynot chose another server for the org default server.

also servers may have http or https. I dont know which it will be. can you take these info to add server functionality too.