the snapshot service is now live.
Integrate the snapshot api to the add new camera feature in the second step of it.

So lets simiplify the second page or second step of the add new camera. lets remove the arrow mark ui and everything. Since we are doing the camera connection test and snapshot creation all within the snapshot api, and since we are not sending multiple responses during the process, we can simplify with a simple loading step with a text saying we are connecting the camera and extracting a snapshot. When the image is available we display it.

Also I forgot to check if in the snapshot api if we have mentioned the bucket name or something and if all the necessary permissions are made