Remove any timeout you have done using signal. No timeouts.
Some critical changes needed are:

So a sample rtsp url is rtsp://user:User@12144@201.76.191.250:8055/cam/realmonitor?channel=1&subtype=0

I need you to add 2 fields called username and password. the values can be user written or when the rtsp url is added you extract the username and password. In this example, its user and 12144. You need to store it appropriately in the db as well. So update the collection as well.